pyproject.toml
[tool.poetry]
name = "madrasvaasi"
version = "0.1.0"
description = ""
authors = ["Srivatsav09 <srivatsav.vijay2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
python-dotenv = "^1.2.1"
django-filter = "^25.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
